Better feedback for backlog when running during a live acquisition
- Periodically check for new tiles to queue in a stage, even if there are already tiles queued
- Include the project tile status worker (dashboard.json reader) in determining the tile status in a plane.

For the latter, it is already determining the sample Z range.  However, by not contributing to the individual tile status in each plane, the tilemap was just showing a blank map for z indices where no tiles have been queued (or if there is just a project and no stages).